ISSUE-23 Separate simulation and scenario for STORE commands
